+++ src/main/java/io/bryansk/icekubit/zhukovcurrencyexchange/dto/ExchangeResponseDto.java
@@ -2,12 +2,14 @@ package io.bryansk.icekubit.zhukovcurrencyexchange.dto;
 
 import io.bryansk.icekubit.zhukovcurrencyexchange.model.Currency;
 
+import java.math.BigDecimal;
+
 public class ExchangeResponseDto {
     private Currency baseCurrency;
     private Currency targetCurrency;
-    private double rate;
-    private double amount;
-    private double convertedAmount;
+    private BigDecimal rate;
+    private BigDecimal amount;
+    private BigDecimal convertedAmount;
 
     public ExchangeResponseDto() {
     }
@@ -28,27 +30,27 @@ public class ExchangeResponseDto {
         this.targetCurrency = targetCurrency;
     }
 
-    public double getRate() {
+    public BigDecimal getRate() {
         return rate;
     }
 
-    public void setRate(double rate) {
+    public void setRate(BigDecimal rate) {
         this.rate = rate;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
 
-    public double getConvertedAmount() {
+    public BigDecimal getConvertedAmount() {
         return convertedAmount;
     }
 
-    public void setConvertedAmount(double convertedAmount) {
+    public void setConvertedAmount(BigDecimal convertedAmount) {
         this.convertedAmount = convertedAmount;
     }
 }
